HBASE-6876 Clean up WARNs and log messages around startup
git-svn-id: https://svn.apache.org/repos/asf/hbase/trunk@1390846 13f79535-47bb-0310-9956-ffa450edef68

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/master/ActiveMasterManager.java
@@ -145,10 +145,11 @@ class ActiveMasterManager extends ZooKeeperListener {
             this.watcher.getMasterAddressZNode(), this.sn)) {
 
           // If we were a backup master before, delete our ZNode from the backup
-          // master directory since we are the active now
-          LOG.info("Deleting ZNode for " + backupZNode + " from backup master directory");
-          ZKUtil.deleteNodeFailSilent(this.watcher, backupZNode);
-
+          // master directory since we are the active now)
+          if (ZKUtil.checkExists(this.watcher, backupZNode) != -1) {
+            LOG.info("Deleting ZNode for " + backupZNode + " from backup master directory");
+            ZKUtil.deleteNodeFailSilent(this.watcher, backupZNode);
+          }
           // Save the znode in a file, this will allow to check if we crash in the launch scripts
           ZNodeClearer.writeMyEphemeralNodeOnDisk(this.sn.toString());
